Revert "[+] Changed Price and salesPrice from Double to BigDecimal"
This reverts commit 9490c8fb8395f6e0fbdce96e908ed705cec64fb0.

+++ src/main/java/com/bitsmilez/productmicroservice/core/domain/model/Product.java
@@ -8,7 +8,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.modelmapper.ModelMapper;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 @Entity
@@ -22,8 +21,8 @@ public class Product {
     @GeneratedValue(strategy = GenerationType.UUID)
     private UUID id;
     private String name;
-    private BigDecimal price;
-    private BigDecimal salesPrice = null;
+    private Double price;
+    private Double salesPrice = null;
     private String img;
     private String description;
     private int stock;
