+++ src/main/java/com/bitsmilez/productmicroservice/core/domain/service/imp/ProductDto.java
@@ -10,7 +10,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.modelmapper.ModelMapper;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -20,8 +19,8 @@ public class ProductDto {
     @GeneratedValue(strategy = GenerationType.UUID)
     private UUID id;
     private String name;
-    private BigDecimal price;
-    private BigDecimal salesPrice;
+    private Double price;
+    private Double salesPrice;
     private String img;
     private String description;
     private int stock;
